First part of layout refactor
