+++ ui/Screen/NotFound.svelte
@@ -1,10 +1,7 @@
 <script>
   import { Flex, Title } from "../DesignSystem/Primitive";
-  import { SidebarLayout } from "../DesignSystem/Component";
 </script>
 
-<SidebarLayout>
-  <Flex align="left">
-    <Title variant="big">Not Found</Title>
-  </Flex>
-</SidebarLayout>
+<Flex align="left">
+  <Title variant="big">Not Found</Title>
+</Flex>
