+++ ui/Screen/Project.svelte
@@ -9,7 +9,6 @@
     AdditionalActionsDropdown,
     HorizontalMenu,
     Remote,
-    SidebarLayout,
     Topbar,
     TrackToggle,
   } from "../DesignSystem/Component";
@@ -111,33 +110,29 @@
   fetch({ id: params.id });
 </script>
 
-<SidebarLayout
-  style="margin: calc(var(--topbar-height)) 0 0 0; min-width: 60em;"
-  dataCy="project-screen">
-  <Remote {store} let:data={project} context="project">
-    <Topbar style="position: fixed; top: 0;">
-      <a slot="left" href={path.projectSource(params.id)} use:link>
-        <!-- TODO(rudolfs): show whether the project is registered under user or org -->
-        <Breadcrumb
-          title={project.metadata.name}
-          user={project.registered}
-          org={project.registered} />
-      </a>
+<Remote {store} let:data={project} context="project">
+  <Topbar style="position: fixed; top: 0;">
+    <a slot="left" href={path.projectSource(params.id)} use:link>
+      <!-- TODO(rudolfs): show whether the project is registered under user or org -->
+      <Breadcrumb
+        title={project.metadata.name}
+        user={project.registered}
+        org={project.registered} />
+    </a>
 
-      <div slot="middle">
-        <HorizontalMenu items={topbarMenuItems(params.id)} />
-      </div>
+    <div slot="middle">
+      <HorizontalMenu items={topbarMenuItems(params.id)} />
+    </div>
 
-      <div slot="right" style="display: flex">
-        <Router routes={menuRoutes} />
-        <TrackToggle style="margin-left: 16px" peerCount="1.3k" />
-        <AdditionalActionsDropdown
-          dataCy="context-menu"
-          style="margin: 0 24px 0 16px"
-          headerTitle={project.shareableEntityIdentifier}
-          menuItems={dropdownMenuItems} />
-      </div>
-    </Topbar>
-    <Router {routes} />
-  </Remote>
-</SidebarLayout>
+    <div slot="right" style="display: flex">
+      <Router routes={menuRoutes} />
+      <TrackToggle style="margin-left: 16px" peerCount="1.3k" />
+      <AdditionalActionsDropdown
+        dataCy="context-menu"
+        style="margin: 0 24px 0 16px"
+        headerTitle={project.shareableEntityIdentifier}
+        menuItems={dropdownMenuItems} />
+    </div>
+  </Topbar>
+  <Router {routes} />
+</Remote>
