+++ ui/Screen/Search.svelte
@@ -1,10 +1,7 @@
 <script>
   import { Flex, Title } from "../DesignSystem/Primitive";
-  import { SidebarLayout } from "../DesignSystem/Component";
 </script>
 
-<SidebarLayout dataCy="page">
-  <Flex align="left">
-    <Title variant="big">Search</Title>
-  </Flex>
-</SidebarLayout>
+<Flex align="left">
+  <Title variant="big">Search</Title>
+</Flex>
