+++ ui/Screen/Settings.svelte
@@ -9,7 +9,7 @@
   import { networkOptions, themeOptions } from "../src/settings.ts";
 
   import { Title, Text, Button } from "../DesignSystem/Primitive";
-  import { SidebarLayout, SegmentedControl } from "../DesignSystem/Component";
+  import SegmentedControl from "../DesignSystem/Component/SegmentedControl.svelte";
 
   const updateNetwork = event =>
     updateRegistry({ ...$settings.registry, network: event.detail });
@@ -45,150 +45,146 @@
   }
 </style>
 
-<SidebarLayout dataCy="page">
-  <div class="container">
-    <Title variant="big">Settings</Title>
+<div class="container">
+  <Title variant="big">Settings</Title>
 
-    <section>
-      <header>
-        <Title variant="large">Version</Title>
-      </header>
-      <div class="section-item">
-        <div class="info">
-          <Text>Version 01.45.02</Text>
-        </div>
-        <div class="action">
-          <Text>There’s a new version of Radicle Upstream</Text>
-          <Button style="margin-left: 16px;">Update to Version 01.45.03</Button>
-        </div>
+  <section>
+    <header>
+      <Title variant="large">Version</Title>
+    </header>
+    <div class="section-item">
+      <div class="info">
+        <Text>Version 01.45.02</Text>
       </div>
-    </section>
+      <div class="action">
+        <Text>There’s a new version of Radicle Upstream</Text>
+        <Button style="margin-left: 16px;">Update to Version 01.45.03</Button>
+      </div>
+    </div>
+  </section>
 
-    <section>
-      <header>
-        <Title variant="large">Appearance</Title>
-      </header>
-      <div class="section-item">
-        <div class="info">
-          <Text variant="medium">Theme</Text>
-        </div>
-        <div class="action">
-          <SegmentedControl
-            active={$settings.appearance.theme}
-            options={themeOptions}
-            on:select={updateTheme} />
-        </div>
+  <section>
+    <header>
+      <Title variant="large">Appearance</Title>
+    </header>
+    <div class="section-item">
+      <div class="info">
+        <Text variant="medium">Theme</Text>
       </div>
-    </section>
+      <div class="action">
+        <SegmentedControl
+          active={$settings.appearance.theme}
+          options={themeOptions}
+          on:select={updateTheme} />
+      </div>
+    </div>
+  </section>
 
-    <section>
-      <header>
-        <Title variant="large">Registry</Title>
-      </header>
-      <div class="section-item">
-        <div class="info">
-          <Text variant="medium">Network</Text>
-        </div>
-        <div class="action">
-          <SegmentedControl
-            active={$settings.registry.network}
-            options={networkOptions}
-            on:select={updateNetwork} />
-        </div>
+  <section>
+    <header>
+      <Title variant="large">Registry</Title>
+    </header>
+    <div class="section-item">
+      <div class="info">
+        <Text variant="medium">Network</Text>
       </div>
-    </section>
+      <div class="action">
+        <SegmentedControl
+          active={$settings.registry.network}
+          options={networkOptions}
+          on:select={updateNetwork} />
+      </div>
+    </div>
+  </section>
 
-    <section>
-      <header>
-        <Title variant="large">Session management</Title>
-      </header>
-      <div class="section-item">
-        <div class="info">
-          <Text variant="medium">Clear local cache</Text>
-          <Text>
-            Removes all locally-stored temporary data from your device.
-          </Text>
-        </div>
-        <div class="action">
-          <Button
-            dataCy="clear-cache-button"
-            variant="outline"
-            on:click={clearCache}>
-            Clear cache
-          </Button>
-        </div>
+  <section>
+    <header>
+      <Title variant="large">Session management</Title>
+    </header>
+    <div class="section-item">
+      <div class="info">
+        <Text variant="medium">Clear local cache</Text>
+        <Text>Removes all locally-stored temporary data from your device.</Text>
       </div>
-      <div class="section-item">
-        <div class="info">
-          <Text variant="medium">Clears all authentication data</Text>
-          <Text>
-            This is similar to how logout works. You will have to create a new
-            identity or restore your existing identity.
-          </Text>
-        </div>
-        <div class="action">
-          <Button
-            dataCy="clear-session-button"
-            variant="outline"
-            on:click={clear}>
-            Clear session
-          </Button>
-        </div>
+      <div class="action">
+        <Button
+          dataCy="clear-cache-button"
+          variant="outline"
+          on:click={clearCache}>
+          Clear cache
+        </Button>
       </div>
-    </section>
+    </div>
+    <div class="section-item">
+      <div class="info">
+        <Text variant="medium">Clears all authentication data</Text>
+        <Text>
+          This is similar to how logout works. You will have to create a new
+          identity or restore your existing identity.
+        </Text>
+      </div>
+      <div class="action">
+        <Button
+          dataCy="clear-session-button"
+          variant="outline"
+          on:click={clear}>
+          Clear session
+        </Button>
+      </div>
+    </div>
+  </section>
 
-    <section>
-      <header>
-        <Title variant="large">Developer tools</Title>
-      </header>
-      <div class="section-item">
-        <div class="info">
-          <Text variant="medium">
-            Unlink all unregistered projects from my profile
-          </Text>
-          <Text>
-            This unlinks your local repositories from Upstream. The local data
-            will remain on your computer.
-          </Text>
-        </div>
-        <div class="action">
-          <Button variant="destructive">Remove</Button>
-        </div>
+  <section>
+    <header>
+      <Title variant="large">Developer tools</Title>
+    </header>
+    <div class="section-item">
+      <div class="info">
+        <Text variant="medium">
+          Unlink all unregistered projects from my profile
+        </Text>
+        <Text>
+          This unlinks your local repositories from Upstream. The local data
+          will remain on your computer.
+        </Text>
       </div>
-    </section>
+      <div class="action">
+        <Button variant="destructive">Remove</Button>
+      </div>
+    </div>
+  </section>
 
-    <section>
-      <header>
-        <Title variant="large">Legal</Title>
-      </header>
-      <div class="section-item">
-        <div class="info">
-          <Text variant="medium">Twemoji</Text>
-          <Text>
-            Copyright 2020 Twitter, Inc and other contributors. Licensed under
-            CC-BY 4.0
-          </Text>
-        </div>
+  <section>
+    <header>
+      <Title variant="large">Legal</Title>
+    </header>
+    <div class="section-item">
+      <div class="info">
+        <Text variant="medium">Twemoji</Text>
+        <Text>
+          Copyright 2020 Twitter, Inc and other contributors. Licensed under
+          CC-BY 4.0
+        </Text>
       </div>
-      <div class="section-item">
-        <div class="info">
-          <Text variant="medium">Inter</Text>
-          <Text>
-            Inter font by Rasmus Andersson licensed under the SIL Open Font
-            License 1.1
-          </Text>
-        </div>
+    </div>
+    <div class="section-item">
+      <div class="info">
+        <Text variant="medium">Inter</Text>
+        <Text>
+          Inter font by Rasmus Andersson licensed under the SIL Open Font
+          License 1.1
+        </Text>
       </div>
-      <div class="section-item">
-        <div class="info">
-          <Text variant="medium">Source Code Pro</Text>
-          <Text>
-            Source Code Pro font by Adobe Fonts distributed under the SIL Open
-            Font License.
-          </Text>
-        </div>
+    </div>
+    <div class="section-item">
+      <div class="info">
+        <Text variant="medium">Source Code Pro</Text>
+        <Text>
+          Source Code Pro font by Adobe Fonts distributed under the SIL Open
+          Font License.
+        </Text>
       </div>
-    </section>
+    </div>
+  </section>
 
-  </div>
-</SidebarLayout>
+</div>
